Added collision detection, can hit enemies and enemies can hit but not hurt player for now

+++ core/src/main/java/com/turmoillift2/handlers/B2DVars.java
@@ -2,7 +2,7 @@ package com.turmoillift2.handlers;
 
 public class B2DVars {
     public static final int PPM = 100;
-    public static final int PLAYER_BIT = 2;
-    public static final int PROJECTILE_BIT = 4;
-    public static final int ENEMY_BIT = 8;
+    public static final short PLAYER_BIT = 2;
+    public static final short PROJECTILE_BIT = 4;
+    public static final short ENEMY_BIT = 8;
 }
